requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.5.0
django-cors-headers==3.3.0
requests==2.28.1
drf-util==1.0.17
drf-yasg==1.21.4
flex==6.13.2
swagger-spec-validator==2.4.1
blinker==1.4
Faker==4.1.0
coverage==6.5.0
django-nose==1.4.7
django-extensions==3.2.1
jsonschema==3.2.0
PyJWT==1.7.1
django-filter==22.1